requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

concepcy==0.1.0
coreferee==1.4.1
matplotlib==3.5.1
networkx==2.7.1
nltk==3.8
pandas==2.0.3
requests==2.32.3
spacy==3.7.5
torch==2.6.0
tqdm==4.66.1
transformers==4.41.2
numpy==1.26.4
openpyxl
neo4j
python-dotenv
boto3==1.36.13
